app/requirements.txt
Django
djangorestframework
numpy
pandas